requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.4.0
calc==1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.6
config==0.5.1
contourpy==1.3.0
cycler==0.12.1
DateTime==5.5
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1
Eel==0.17.0
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.0.3
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
Greeting==1.0.1
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.context==5.3.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.4.0
MouseInfo==0.1.3
multidict==6.0.5
narwhals==1.14.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
newsapi-python==0.2.7
nltk==3.8.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.0
oauthlib==3.2.2
openai==0.28.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
playsound==1.3.0
plyer==2.1.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
pyjokes==0.8.3
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
ratelim==0.1.6
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.23.1
Send2Trash==1.8.3
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
South==1.0.2
SpeechRecognition==3.10.4
speedtest==0.0.1
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.40.2
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
vectorizer==0.3.0
Voice==0.1.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.44.0
whichcraft==0.6.1
widgetsnbextension==4.0.13
wikipedia==1.4.0
wolframalpha==5.1.3
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.4
zope.event==5.0
zope.interface==7.0.1


